Regex module rename.
