+++ google/cloud/forseti/common/util/regular_exp.py
