+++ google/cloud/forseti/scanner/audit/buckets_rules_engine.py
@@ -21,7 +21,7 @@ import re
 from google.cloud.forseti.common.gcp_type.bucket_access_controls import (
     BucketAccessControls)
 from google.cloud.forseti.common.util import logger
-from google.cloud.forseti.common.util.regular_expressions import escape_and_globify
+from google.cloud.forseti.common.util.regular_exp import escape_and_globify
 from google.cloud.forseti.scanner.audit import base_rules_engine as bre
 from google.cloud.forseti.scanner.audit import errors as audit_errors
