+++ google/cloud/forseti/scanner/audit/instance_network_interface_rules_engine.py
@@ -17,7 +17,7 @@ from collections import namedtuple
 import itertools
 import re
 
-from google.cloud.forseti.common.util.regular_expressions import escape_and_globify
+from google.cloud.forseti.common.util.regular_exp import escape_and_globify
 from google.cloud.forseti.common.util import logger
 from google.cloud.forseti.scanner.audit import base_rules_engine as bre
 from google.cloud.forseti.scanner.audit import errors as audit_errors
